Instantiate store instantly for global availability
