+++ packages/scrabble-solver/src/state/index.ts
@@ -1,7 +1,7 @@
 export * from './actions';
-export { default as createAppStore } from './createAppStore';
 export { default as localStorage } from './localStorage';
 export * from './selectors';
 export * from './slices';
+export { default as store } from './store';
 export { default as useTranslate } from './useTranslate';
 export { default as useTypedSelector } from './useTypedSelector';
